+++ Front_end/areum/src/pages/Group/Group.css
@@ -68,4 +68,3 @@
     width: 100%;
     padding: 10px;
   }
-
