+++ Front_end/areum/src/pages/Group/Group.js
@@ -2,6 +2,7 @@ import React, { useRef,useState,useEffect } from "react";
 import axios from 'axios';
 import { Link } from "react-router-dom";
 import "./Group.css";
+import LoginBox from "../../components/LoginBox";
 import CustomHeader from "../../components/CustomHeader";
 
 function Group() {
@@ -9,9 +10,6 @@ function Group() {
   const [posts,setPosts]=useState([]);
   const [searchText, setSearchText] = useState("");
   const [photo, setPhoto] = useState(null);
-
-  const user=JSON.parse(localStorage.getItem('user'));
-  const userImg=localStorage.getItem('img');
 
   const token=localStorage.getItem('token');
       const baseAxios=axios.create();
@@ -55,26 +53,7 @@ function Group() {
 
 <div className="container">
 <div className="login-box">
-<div className="profile">
-  <div className="profile-info">
-    <div className="profile-top">
-      <div className="profile-circle">
-        {userImg && <img className="profile-photo" src={userImg} alt="Profile" />}
-      </div>
-      <div className="name-container">
-        <span>{user.name} 님 </span>
-      </div>
-    </div>
-    <div className="department-grade-container">
-      <p className="department-grade">
-        <span className="department-text">
-          {user.major} 학과
-        </span>
-      </p>
-    </div>
-    <hr className="content-divider2" />
-  </div>
-</div>
+<LoginBox/>
 
   <div className="MP-button">
   <Link to ="/MyPage">
@@ -107,7 +86,7 @@ function Group() {
               <hr className="content-divider" />
             </div>
 
-            <div className="contents">
+            <div className="group-contents">
               {posts.map((page) => (
                 <React.Fragment key={page.id}>
                   <a href={`/PostDetail/${page.id}`} style={{ textDecoration: "none" }}>
